+++ csatar-plugins/knowledgerepository/updates/SeederData.php
@@ -2,6 +2,7 @@
 
 use Csatar\KnowledgeRepository\Models\AccidentRiskLevel;
 use Csatar\KnowledgeRepository\Models\GameDevelopmentGoal;
+use Csatar\KnowledgeRepository\Models\MethodologyType;
 use Csatar\KnowledgeRepository\Models\Tool;
 use Csatar\KnowledgeRepository\Models\HeadCount;
 use Csatar\KnowledgeRepository\Models\Duration;
@@ -260,6 +261,20 @@ class SeederData extends Seeder
                 'sort_order' => 6,
             ],
         ],
+        'methodologyTypes' => [
+            [
+                'name' => 'új (eddig az őrs még sosem találkozott az átadott anyaggal)',
+                'sort_order' => 1
+            ],
+            [
+                'name' => 'régi/ismétlés (ismétlés, az őrs már találkozott a megnevezett anyaggal)',
+                'sort_order' => 2
+            ],
+            [
+                'name' => 'mindkettő',
+                'sort_order' => 3
+            ],
+        ]
     ];
 
     public function run()
@@ -329,6 +344,16 @@ class SeederData extends Seeder
             ]);
             $gameType->sort_order = $gameTypeData['sort_order'];
             $gameType->save();
+        }
+
+        // Methodology Types
+
+        foreach ($this::DATA['methodologyTypes'] as $methodologyTypeData) {
+            $methodologyType = MethodologyType::firstOrNew([
+                'name' => $methodologyTypeData['name'],
+            ]);
+            $methodologyType->sort_order = $methodologyTypeData['sort_order'];
+            $methodologyType->save();
         }
     }
 }